add jdk utils
